requirements.txt
# Deep Learning
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Face Recognition
facenet-pytorch==2.6.0

# Computer Vision
opencv-python>=4.8.0

# Web Frameworks
streamlit>=1.30.0
flask>=3.0.0

# Data Science
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilities
scipy>=1.11.0
Pillow>=10.0.0

# Jupyter
ipykernel>=6.25.0
jupyter>=1.0.0